requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
markdown==3.7
django-filter==24.3
cryptography==44.0.0
django-redis==5.4.0
redis==5.2.1

boto3==1.35.97
django-storages==1.14.4
Pillow==11.1.0
django-minio-storage==0.5.7
django-imagekit==5.0.0

celery==5.4.0
django-celery-results==2.5.1
django_celery_beat==2.7.0
mongoengine==0.29.1
